refactor: 예외가 존재하는 쿼리문을 맨 위로 올림
