+++ src/main/java/wooteco/subway/admin/domain/subwayShortestPath/SubwayShortestPath.java
@@ -1,6 +1,8 @@
-package wooteco.subway.admin.domain;
+package wooteco.subway.admin.domain.subwayShortestPath;
 
 import java.util.List;
+
+import wooteco.subway.admin.domain.Station;
 
 public interface SubwayShortestPath {
 	List<Station> getVertexList();
